+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 - **Feature: Support Ruby 3.4.0**
 
   The agent now supports Ruby 3.4.0. We've made incremental changes throughout the preview stage to reach compatibility. This release includes an update to the Thread Profiler for compatibility with Ruby 3.4.0's new backtrace format. [Issue#2992](https://github.com/newrelic/newrelic-ruby-agent/issues/2992) [PR#2997](https://github.com/newrelic/newrelic-ruby-agent/pull/2997)
+  
+- **Feature: Add instrumentation for aws-sdk-firehose**
+
+  The agent now has instrumentation for the [aws-sdk-firehose](https://rubygems.org/gems/aws-sdk-firehose) gem. [PR#2973](https://github.com/newrelic/newrelic-ruby-agent/pull/2973)
 
 - **Feature: Kubernetes APM auto-attach - new agent version precedent**